@@ -28,11 +28,11 @@ android {
         exclude 'lib/x86/libruntimecore_java.so'
 //        exclude 'lib/armeabi-v7a/libruntimecore_java.so'
     }
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.app.leon.sellabfa"
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -85,19 +85,29 @@ android {
 //                multiDexEnabled true
 //            }
 //        }
+        compileOptions {
+            sourceCompatibility JavaVersion.VERSION_1_8
+            targetCompatibility JavaVersion.VERSION_1_8}
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8}
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
+    //noinspection GradleCompatible
     implementation 'com.android.support:appcompat-v7:28.0.0'
+    //noinspection GradleCompatible
     implementation 'com.android.support:design:28.0.0'
+    //noinspection GradleCompatible
     implementation 'com.android.support:cardview-v7:28.0.0'
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
     implementation 'com.android.support:multidex:1.0.3'
+    //noinspection GradleCompatible
     implementation 'com.android.support:support-v4:28.0.0'
     implementation 'com.jakewharton:butterknife:10.1.0'
     implementation 'com.jakewharton:butterknife-compiler:10.1.0'
@@ -112,4 +122,7 @@ dependencies {
     implementation 'com.facebook.shimmer:shimmer:0.4.0@aar'
     implementation 'com.esri.arcgisruntime:arcgis-android:100.2.1'
     implementation 'com.google.android.gms:play-services-location:17.0.0'
+
+    implementation "com.android.support:support-annotations:28.0.0"
+    implementation 'com.jakewharton.timber:timber:4.7.1'
 }
