@@ -1,17 +1,10 @@
 package com.app.leon.sellabfa.BaseItem;
 
-
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Typeface;
 import android.os.Bundle;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.Gravity;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -24,6 +17,12 @@ import android.widget.EditText;
 import android.widget.ListView;
 import android.widget.TextView;
 
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+
 import com.app.leon.sellabfa.Activities.DownloadActivity;
 import com.app.leon.sellabfa.Activities.ReadActivity;
 import com.app.leon.sellabfa.Activities.UploadActivity;
@@ -31,6 +30,7 @@ import com.app.leon.sellabfa.Adapters.NavigationCustomAdapter;
 import com.app.leon.sellabfa.Models.ViewModels.UiElementInActivity;
 import com.app.leon.sellabfa.R;
 import com.app.leon.sellabfa.Utils.FontManager;
+import com.google.android.material.navigation.NavigationView;
 
 import java.util.ArrayList;
 import java.util.List;
